requirements.txt
flask
pymongo
faiss-cpu
numpy
tqdm
